requirements.txt
aiohttp==3.8.4
chat_exporter==2.6.1
discord.py==2.3.2
jishaku==2.5.1
lavalink==4.0.6
lxml==4.9.3
psutil==5.9.4
python_dateutil==2.8.2
pytz==2022.7
typing_extensions==4.8.0
yarl==1.8.2
